requirements.txt
shortuuid==0.4.3
mustache==0.1.4
configobj==5.0.6
wikipedia==1.4.0
requests==2.8.1
pyOpenSSL==16.0.0
ndg-httpsclient==0.4.0
pyasn1==0.1.9
gTTS==1.0.7
backports.ssl-match-hostname==3.4.0.2
certifi==2016.2.28
PyAudio==0.2.8
pyee==1.0.1
SpeechRecognition==3.1.3
tornado==4.2.1
websocket-client==0.32.0
pyowm==2.2.1
wolframalpha==1.4
futures==3.0.3
requests-futures==0.9.5
astral==0.9
tzlocal==1.2
parsedatetime==1.5
pdoc==0.3.2
pyyaml==3.11
feedparser==5.2.1
pyalsaaudio==0.8.2
xmlrunner==1.7.7
pyserial==3.0
netifaces==0.10.4
pyjokes==0.5.0
psutil==4.1.0
pep8==1.7.0
multi_key_dict==2.0.3
pocketsphinx==0.1.0
python-pandora==0.10
git+https://github.com/clusterfudge/adapt.git@v0.2.7#egg=adapt-parser
